Move types into index.ts and remove types.ts
